Wokrs fine now

+++ NicE-Buoy/src/calculate.h
@@ -5,7 +5,7 @@ extern double errSum;
 
 void initCalculate(void);
 void setparameters(int *minOfsetDist, int *maxOfsetDist, int *minSpeed, int *maxSpeed);
-int CalcDocSpeed(double tgdistance);
+double CalcDocSpeed(double tgdistance);
 void adjustPositionDirDist(int dir, double dist, double *tglatitude, double *tglongitude);
 void CalcRemoteRudderBuoy(double magheading, float tgheading, int speed, int *bb, int *sb);
 bool CalcRudderBuoy(double magheading, float tgheading,double tdistance, int speed, int *bb, int *sb);
